requirements.txt
streamlit>=1.30.0
python-pptx>=0.6.21
openpyxl>=3.1.0
python-docx>=1.1.0
google-generativeai>=0.8.0
groq>=0.9.0
cohere>=5.0.0
